% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_to_cid.R
\name{post_to_cid}
\alias{post_to_cid}
\title{POST Information to Retrieve PubChem CIDs}
\source{
https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest-tutorial
}
\usage{
post_to_cid(x, format, domain = "compound", json = FALSE)
}
\arguments{
\item{x}{(Character.) Input to the search, must match with the format.}

\item{format}{(Character.) Format type of the input. This is not case
sensitive. Must be lower case! See Details for supported formats.}

\item{domain}{(Character.) Must be either \code{"compound"} (default) or
\code{"substance"}.}

\item{json}{(Logical.) Should the result be returned as JSON? Defaults to
\code{FALSE}.}
}
\value{
Returns an integer (or possibly vector thereof) or a character
  string, depending on the value of \code{json}.
}
\description{
This function performs a query to retrieve PubChem CIDs for
  different inputs.
}
\details{
The function performs a sanity check on the provided inputs and
  then performs a query. If successful, an integer vector with the available
  PubChem CID(s) will be returned for the \code{"compound"} domain, or a
  list of PubChem SID(s) and CID(s) for the \code{"substance"} domain.

  Supported formats include \code{"InChI"}, \code{"InChIKey"},
  \code{"Name"}, \code{"SDF"}, and \code{"SMILES"}.
}
\examples{
\dontrun{
name <- "Atrazine"
post_to_cid(name, format = "Name")
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
