% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_to_property.R
\name{post_to_property}
\alias{post_to_property}
\title{POST PubChem CIDs to Retrieve Their Properties}
\source{
https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest-tutorial
}
\usage{
post_to_property(cid, property = "all", json = FALSE)
}
\arguments{
\item{cid}{(Integer) PubChem CID(s) as single integer or a vector of
integers.}

\item{property}{(Character) Properties to retrieve as single character or
vector of characters. This is case-sensitive! Defaults to \code{"all"}.
See Details for all available properties.}

\item{json}{(Logical) Should the result be returned as JSON? Defaults to
\code{FALSE}.}
}
\value{
Returns a data frame or a list, depending on the value of
  \code{extract}.
}
\description{
This function performs a query to retrieve properties for
  PubChem CIDs.
}
\details{
The function performs a sanity check on the provided PubChem CIDs
  and properties and then performs a query. If successful, a list with the
  available properties will be returned.

  Supported properties include \code{"MolecularFormula"},
  \code{"MolecularWeight"}, \code{"CanonicalSMILES"},
  \code{"IsomericSMILES"}, \code{"InChI"}, \code{"InChIKey"},
  \code{"IUPACName"}, \code{"Title"}, \code{"XLogP"}, \code{"ExactMass"},
  \code{"MonoisotopicMass"}, \code{"TPSA"}, \code{"Complexity"},
  \code{"Charge"}, \code{"HBondDonorCount"}, \code{"HBondAcceptorCount"},
  \code{"RotatableBondCount"}, \code{"HeavyAtomCount"},
  \code{"IsotopeAtomCount"}, \code{"AtomStereoCount"},
  \code{"DefinedAtomStereoCount"}, \code{"UndefinedAtomStereoCount"},
  \code{"BondStereoCount"}, \code{"DefinedBondStereoCount"},
  \code{"UndefinedBondStereoCount"}, \code{"CovalentUnitCount"},
  \code{"Volume3D"}, \code{"XStericQuadrupole3D"},
  \code{"YStericQuadrupole3D"}, \code{"ZStericQuadrupole3D"},
  \code{"FeatureCount3D"}, \code{"FeatureAcceptorCount3D"},
  \code{"FeatureDonorCount3D"}, \code{"FeatureAnionCount3D"},
  \code{"FeatureCationCount3D"}, \code{"FeatureRingCount3D"},
  \code{"FeatureHydrophobeCount3D"}, \code{"ConformerModelRMSD3D"},
  \code{"EffectiveRotorCount3D"}, \code{"ConformerCount3D"}, and
  \code{"Fingerprint2D"}. The default, \code{"all"}, retrieves all listed
  properties.
}
\examples{
\dontrun{
cid <- 2244
post_to_property(cid)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
