% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_to_reaction.R
\name{post_to_reaction}
\alias{post_to_reaction}
\title{POST a PubChem CID to Retrieve Its Known Reaction Products}
\usage{
post_to_reaction(cid, collection, json = FALSE)
}
\arguments{
\item{cid}{(Integer) PubChem CID as single integer.}

\item{collection}{(Character) Which reaction collection should be returned?
Must be one of \code{"rhea"} or \code{"pathwayreaction"}.}

\item{json}{(Logical) Should the result be returned as JSON? Defaults to
\code{FALSE}.}
}
\value{
Returns an data frame or a character string, depending on the value
  of \code{json}.
}
\description{
This function performs a query to retrieve known reaction
  products for a PubChem CID.
}
\details{
The function performs a sanity check on the provided PubChem CID
  and collection, then performs a query. If successful, a data frame with
  the available reaction products will be returned.
}
\examples{
\dontrun{
cid <- 2256
collection <- "rhea"
post_to_reaction(cid, collection)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
