% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_to_transformation.R
\name{post_to_transformation}
\alias{post_to_transformation}
\title{POST a PubChem CID to Retrieve Its Transformation Products}
\usage{
post_to_transformation(cid, json = FALSE)
}
\arguments{
\item{cid}{(Integer) PubChem CID as single integer.}

\item{json}{(Logical) Should the result be returned as JSON? Defaults to
\code{FALSE}.}
}
\value{
Returns an data frame or a character string, depending on the value
  of \code{json}.
}
\description{
This function performs a query to retrieve known transformation
  products for a PubChem CID.
}
\details{
The function performs a sanity check on the provided PubChem CID
  and then performs a query. If successful, a data frame with the available
  transformation products will be returned.
}
\examples{
\dontrun{
cid <- 2256
post_to_transformation(cid)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
