% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_to_use.R
\name{post_to_use}
\alias{post_to_use}
\title{POST a PubChem CID to Retrieve Its Use Categories}
\usage{
post_to_use(cid, json = FALSE)
}
\arguments{
\item{cid}{(Integer) PubChem CID as single integer.}

\item{json}{(Logical) Should the result be returned as JSON? Defaults to
\code{FALSE}.}
}
\value{
Returns an data frame or a character string, depending on the value
  of \code{json}.
}
\description{
This function performs a query to retrieve known use categories
  for a PubChem CID.
}
\details{
The function performs a sanity check on the provided PubChem CID
  and then performs a query. If successful, a data frame with the available
  use categories will be returned.
}
\examples{
\dontrun{
cid <- 2256L
post_to_use(cid)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
