% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_to_cids.R
\name{post_to_cid}
\alias{post_to_cid}
\title{POST Information to Retrieve PubChem CIDs}
\source{
https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest-tutorial
}
\usage{
post_to_cid(x, format, json = FALSE)
}
\arguments{
\item{x}{(Character) Input to the search, must match with the format.}

\item{format}{(Character) Format type of the input. This is not case
sensitive. Must be lower case! See Details for supported formats.}

\item{json}{(Logical) Should the result be returned as JSON? Defaults to
\code{FALSE}.}
}
\value{
Returns a list.
}
\description{
This function performs a query to retrieve PubChem CIDs for
  different inputs.
}
\details{
The function performs a sanity check on the provided inputs and
  then performs a query. If successful, a list with the available
  PubChem CIDs will be returned.

  Supported formats include \code{"InChI"}, \code{"InChIKey"},
  \code{"Name"}, \code{"SDF"}, and \code{"SMILES"}.
}
\examples{
\dontrun{
x <- "Aspirin"
post_to_cid(x, format = "name")
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
